+++ agixt/ApiClient.py
@@ -1,6 +1,5 @@
 import os
 import jwt
-import subprocess
 from agixtsdk import AGiXTSDK
 from dotenv import load_dotenv
 from fastapi import Header, HTTPException
@@ -64,13 +63,3 @@ def verify_api_key(authorization: str = Header(None)):
             raise HTTPException(status_code=401, detail="Invalid API Key")
     else:
         return "USER"
-
-
-def is_cuda():
-    try:
-        result = subprocess.run(
-            ["nvidia-smi"], stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True
-        )
-        return "NVIDIA-SMI" in result.stdout
-    except:
-        return False
